% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{contrX2}
\alias{contrX2}
\title{Tableau de contributions au khi-2}
\usage{
contrX2(data)
}
\arguments{
\item{data}{: valeur d'origine}
}
\value{
contribution au khi-2 de chaque observation
}
\description{
Tableau de contributions au khi-2
}

