% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{d2}
\alias{d2}
\title{Le test d'independance}
\usage{
d2(data)
}
\arguments{
\item{data}{: valeur d'origine}
}
\value{
la valeur de d2, la valeur critique et la conclusion de test
}
\description{
Pour mesurer de la liasion entre deux variables qualitatives, on realiser un test khi-2. Cette statistique est la r¨¦alisation d'une variable al¨¦atoire d2 de la loi khi-2 avec degre de libert¨¦ (r-1)(c-1)
}

