% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{gcol}
\alias{gcol}
\title{Centre gravite de colonnes}
\usage{
gcol(data)
}
\arguments{
\item{data:}{valeur d'origine}
}
\value{
un vector de centre de gravite
}
\description{
Le centre de gravite de profils-lignes affect¨¦s avec des poids(frequences)
}

