% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{grow}
\alias{grow}
\title{Centre gravite de lignes}
\usage{
grow(data)
}
\arguments{
\item{data:}{valeur d'origine}
}
\value{
un vector de centre de gravite
}
\description{
Le centre de gravite de profils-colonnes affect¨¦s avec des poids(frequences)
}

