% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{tabFreq}
\alias{tabFreq}
\title{Tableau de frequences}
\usage{
tabFreq(data)
}
\arguments{
\item{data:}{valeur d'origine}
}
\value{
un tableau de fr¨¦quences
}
\description{
Le tableau de fr¨¦quences est construit en divisant le nombre observ¨¦ de la variable par le nombre total d'observation
}

