% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{tabProColonne}
\alias{tabProColonne}
\title{Tableau des profils-colonnes}
\usage{
tabProColonne(data)
}
\arguments{
\item{data:}{valeur d'origine}
}
\value{
un tableau des profils-colonnes
}
\description{
Le tableau des profils en colonnes est construit en divisant l'effectif de chaque case par le total de la colonne correspondante
}

