% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{tabProLigne}
\alias{tabProLigne}
\title{Tableau des profils-lignes}
\usage{
tabProLigne(data)
}
\arguments{
\item{data:}{valeur d'origine}
}
\value{
un tableau des profils-lignes
}
\description{
Le tableau des profils en lignes est construit en divisant l'effectif de chaque case par le total de la ligne correspondante
}

