% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{tabValTheo}
\alias{tabValTheo}
\title{Tableau des valeurs theoriques}
\usage{
tabValTheo(data)
}
\arguments{
\item{data:}{data frame construite par importer les data ¨¤ traiter}
}
\value{
un tableau des valeurs th¨¦oriques
}
\description{
Le tableau des valeurs th¨¦oriques est construit en divisant le total de ligne fois le total de colonne par le total de tableau
}

