% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysed.R
\name{total}
\alias{total}
\title{Totaux}
\usage{
total(data)
}
\arguments{
\item{data:}{data frame construite par importer les data ид traiter}
}
\value{
un tableau de data d'origine avec les totaux lignes et colonnes
}
\description{
Ajouter la colonne et la ligne totaux
}

